+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2024-09-29
+
+- [ADDED] Upgraded to latest version of OpenAI API
+- [FIXED] Unit Tests
+- [FIXED] Bug in saving generated image for use with unit tests
+
 ## [1.7.0] - 2024-09-08
 
 - [ADDED] Safety GitHub Action workflow to check Python dependencies for known security vulnerabilities.